tg-ui/core-bootstrap: migrate HolonsBotCore + ServiceContainer to TS
Convert the bootstrap modules of @holons/telegram-ui to TypeScript so the
remaining handler migration units (Phase B 17-19) can plug into a typed
surface. Behavior is unchanged — this is a typing pass.

- Rename HolonsBotCore.js, ServiceContainer.js, core/index.js to .ts and
  add typed interfaces (ServiceFactory, ServiceMap, PostInitHookMap,
  TelebotLike, DatabaseLike, GunLike, etc.) using `import type` from
  Telegraf to avoid runtime imports.
- Tighten tsconfig.json: noEmit:false, outDir:./dist, declaration:false
  (re-enable once handlers are TS), include only the migrated files.
- Add tsx, typescript, @types/node devDeps and switch dev/start scripts
  to tsx so the JS entrypoint can transparently resolve the new TS
  modules via existing `.js` import specifiers.
- Add `errorMessage()` helper (exported from ServiceContainer) and use
  it to dedupe nine inline `error instanceof Error ? ... : String(...)`
  patterns introduced during the conversion.
- Ignore dist/ and *.tsbuildinfo.
