+++ packages/telegram-ui/tsconfig.json
@@ -2,12 +2,22 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./src",
+    "rootDir": ".",
     "allowJs": true,
     "checkJs": false,
-    "noEmit": true,
+    "noEmit": false,
+    "emitDeclarationOnly": false,
+    // Declarations cannot be safely emitted while most of the package is still
+    // JS — the inferred Telegraf types from JS modules are not portable.
+    // Re-enable once handler modules (units 17-19) are migrated to TS.
+    "declaration": false,
+    "declarationMap": false,
     "types": ["node"]
   },
-  "include": ["src/**/*"],
-  "exclude": ["dist", "node_modules", "**/*.test.js", "**/*.spec.js"]
+  "include": [
+    "core/HolonsBotCore.ts",
+    "core/ServiceContainer.ts",
+    "core/index.ts"
+  ],
+  "exclude": ["dist", "node_modules", "**/*.test.js", "**/*.spec.js", "**/*.test.ts", "**/*.spec.ts"]
 }
